Player class abstraction
Spent a long time (at least two hours) to figure out how to do this. There's probably a better way but I'm surprised enough that it works.

My intention was to allow a real player and an AI player to inherit from the same base class, using virtual/abstract methods. However, I didn't expect abstraction to be so different in C++ than C#.